Refactored - Introduced fields in place of methods. Inlined methods. Moved files to right namespace. Changed namesspace (directory name Api -> Integration)
