+++ src/Tests/CaptainHook.Api.Tests/Integration/ControllerTestBase.cs
@@ -1,10 +1,8 @@
-﻿using CaptainHook.Api.Client;
+﻿using System;
+using CaptainHook.Api.Client;
 using CaptainHook.Api.Tests.Config;
-using System;
-using System.Collections.Generic;
-using System.Text;
 
-namespace CaptainHook.Api.Tests.Api
+namespace CaptainHook.Api.Tests.Integration
 {
     public abstract class ControllerTestBase : IDisposable
     {
